# Python minor performance improvements and simplifications