--- a/.gitignore
+++ b/.gitignore
@@ -8,6 +8,8 @@ related_*.json
 .DS_Store
 .vscode/settings.json
 venv/
+.idea/
+python/*.py.pclprof
 crystal/lib/
 crystal/.shards/
 *.dwarf