--- a/.gitignore
+++ b/.gitignore
@@ -8,6 +8,7 @@ related_*.json
 .DS_Store
 .vscode/settings.json
 venv/
+__pycache__/
 .idea/
 python/*.py.pclprof
 crystal/lib/