--- a/python/related.py
+++ b/python/related.py
@@ -1,9 +1,8 @@
 from timing import lap, finish
 
 lap()
-import heapq
 import json
-from collections import Counter
+from collections import Counter, defaultdict
 
 
 def main():
@@ -12,28 +11,20 @@ def main():
         posts = json.load(f)
     lap()
 
-    tag_map = {}
+    tag_map = defaultdict(list)
     for idx, post in enumerate(posts):
         for tag in post["tags"]:
-            if tag not in tag_map:
-                tag_map[tag] = []
             tag_map[tag].append(idx)
 
     all_related_posts = []
     for this_post_idx, post in enumerate(posts):
-        related_posts_list = Counter((p for tag in post["tags"] for p in tag_map[tag]))
-        related_posts_list[this_post_idx] = 0
-
-        top_posts = [
-            {k: posts[p][k] for k in ("_id", "title", "tags")}
-            for p in heapq.nlargest(5, related_posts_list, key=related_posts_list.get)
-        ]
-
+        relation_count = Counter((p for tag in post["tags"] for p in tag_map[tag]))
+        relation_count[this_post_idx] = 0
         all_related_posts.append(
             {
                 "_id": post["_id"],
                 "tags": post["tags"],
-                "related": top_posts,
+                "related": [posts[p] for p, _ in relation_count.most_common(5)],
             }
         )