--- a/python/related_np.py
+++ b/python/related_np.py
@@ -1,7 +1,7 @@
 from timing import lap, finish
 
 lap()
-import json
+import ujson as json
 import numpy as np