--- a/python/requirements.txt
+++ b/python/requirements.txt
@@ -1,1 +1,2 @@
 numpy
+ujson