--- a/python/timing.py
+++ b/python/timing.py
@@ -1,3 +1,4 @@
+import os
 import time
 
 perf = []
@@ -10,8 +11,13 @@ def lap():
 
 
 def finish():
-    data = "\t".join(f"{b - a}" for a, b in zip(perf, perf[1:]))
-    print(f"perf_counter\t{data}")
-    data = "\t".join(f"{b - a}" for a, b in zip(proc, proc[1:]))
-    print(f"process_time\t{data}")
-
+    detailed = os.environ.get("DETAILED_PYTHON_PERF") == "1"
+    if detailed:
+        data = "\t".join(f"{b - a}" for a, b in zip(perf, perf[1:]))
+        print(f"perf_counter\t{data}")
+        data = "\t".join(f"{b - a}" for a, b in zip(proc, proc[1:]))
+        print(f"process_time\t{data}")
+    else:
+        end = perf[-2]  # before final output
+        start = perf[-3]  # before process and output
+        print(f"Processing time (w/o IO): {end - start:.3f}s")