--- a/run.sh
+++ b/run.sh
@@ -94,12 +94,17 @@ run_rust_rayon() {
 run_python() {
     echo "Running Python" &&
         cd ./python &&
+        if [ ! -d "venv" ]; then
+            python3 -m venv venv
+        fi
+    source venv/bin/activate &&
+        pip freeze | grep ujson || pip install -r requirements.txt &&
         if [ $HYPER == 1 ]; then
             capture "Python" hyperfine -r 5 --show-output "python3 ./related.py"
         else
             command time -f '%es %Mk' python3 ./related.py
         fi
-
+    deactivate
     check_output "related_posts_python.json"
 
 }
@@ -111,7 +116,7 @@ run_python_np() {
             python3 -m venv venv
         fi
     source venv/bin/activate &&
-        pip freeze | grep numpy || pip install -r requirements.txt &&
+        (pip freeze | grep numpy && pip freeze | grep ujson) || pip install -r requirements.txt &&
         if [ $HYPER == 1 ]; then
             capture "Numpy" hyperfine -r 5 --show-output "python3 ./related_np.py"
         else