1. report more detailed performance info (used [here](https://github.com/fizmat/related_post_gen_python_timing/blob/main/main.ipynb) to measure the rest of the changes)
2. run: install everything from requirements.txt (in case more libraries will be used in the future);
3. run: also remove venv in cleanup;
4. python: use defaultdict. Just convenience, no performance impact;
5. python: remove heapq. Counter can find top-k itself. Just convenience, no performance impact;
6. both: use ujson as a transparent replacement for json. 90 ms faster output;
7. numpy: generate unique tags using python set. 20 ms faster than filling an ndarray and then calling `.unique()`;
8. numpy: iterate over post.tags dict instead of iterating over all possible tag ids in the inner loop. 20ms faster;
9. use uint8 as matrix dtype: 20ms faster than uint16;

Interesting result in c9414160a56d5122a8ccb738a8675d394fdfa987: using `np.argpartition()` is 30ms slower than `np.argsort()`, even though partition does less (it does not have to fully sort the array). I think the key is `kind="stable"`. There's no algorithm selection for argpartition yet, and stablesort seems to be very fast on this specific dataset. Are related posts almost-sorted by relatedness?