@fizmat 2023-10-01
Ooh, "stable" argsort is probably using radix sort in this case, which should be super fast given relatedness has very few possible values. https://numpy.org/doc/stable/reference/generated/numpy.argsort.html#numpy.argsort