@jinyus 2023-10-01
Hey, thanks for your efforts but we need the timing to be compatible with the rest of the solutions because the benchmark results are automatically generated. The rest of your changes looks good.

Let me know if you're going to modify the pull request or I can do it.