@fizmat 2023-10-01
Oh, I didn't know. What does the benchmark actually collect, the "Processing time (w/o IO): ..." lines?