@jinyus 2023-10-01
Yes. I am also trying to limit the run time of the workflow because the free tier is limited.

[link to regex](https://github.com/jinyus/related_post_gen/blob/30c4b3067df562e35e0baa62372af90107e61ab3/extract_results.dart#L7)