@fizmat 2023-10-01
Changed. Now it only prints the detailed table if env DETAILED_PYTHON_PERF="1", otherwise behaves as usual.
Also now using `time` functions that return int nanoseconds instead of float seconds (it will not actually improve precision given the current scale of ~ second per run, but why not).