@jinyus 2023-10-01
Processing times is a bit slower on my machine. ujson sped up I/O though. What are the results on your machine?

new:

```
Benchmark 1: python3 ./related.py
Processing time (w/o IO): 1.425s
Processing time (w/o IO): 1.413s
Processing time (w/o IO): 1.421s
Processing time (w/o IO): 1.409s
Processing time (w/o IO): 1.415s
  Time (mean ± σ):      1.485 s ±  0.007 s    [User: 1.465 s, System: 0.018 s]
  Range (min … max):    1.478 s …  1.492 s    5 runs

```

old:

```Benchmark 1: python3 ./related.py
Processing time (w/o IO): 1.366s
Processing time (w/o IO): 1.353s
Processing time (w/o IO): 1.330s
Processing time (w/o IO): 1.349s
Processing time (w/o IO): 1.310s
  Time (mean ± σ):      1.516 s ±  0.023 s    [User: 1.492 s, System: 0.024 s]
  Range (min … max):    1.489 s …  1.546 s    5 runs
```