@fizmat 2023-10-01
old:
```
Benchmark 1: python3 ./related.py
Processing time (w/o IO): 1.344s
Processing time (w/o IO): 1.351s
Processing time (w/o IO): 1.350s
Processing time (w/o IO): 1.342s
Processing time (w/o IO): 1.350s
  Time (mean ± σ):      1.497 s ±  0.005 s    [User: 1.477 s, System: 0.014 s]
  Range (min … max):    1.493 s …  1.503 s    5 runs
```

new:
```
Benchmark 1: python3 ./related.py
Processing time (w/o IO): 1.473s
Processing time (w/o IO): 1.465s
Processing time (w/o IO): 1.461s
Processing time (w/o IO): 1.458s
Processing time (w/o IO): 1.464s
  Time (mean ± σ):      1.508 s ±  0.027 s    [User: 1.486 s, System: 0.012 s]
  Range (min … max):    1.491 s …  1.555 s    5 runs
```

I checked, and now commit `minor pythonizations: deafultdict, no need for heapq` slows processing down by almost 120ms. Strange.