@fizmat 2023-10-01
I think I know why I didn't see a difference before: version differences!

The latest result, where it's 120ms slower are in a py=3.9 venv, but the jupyter notebook with performance plots I generated in py=3.11. Probably either counter or defaultdict got optimized between 3.9 and 3.11.