@fizmat 2023-10-01
Yup, plot in 3.9:
![image](https://github.com/jinyus/related_post_gen/assets/7406466/04a1904f-0d3c-46b0-a0b1-426ee3fc1d0d)
Plot in 3.11:
![image](https://github.com/jinyus/related_post_gen/assets/7406466/46514400-34d0-4d60-b336-908507beb870)