@jinyus 2023-10-01
I have 3.11.5 on my machine. It's slower on the vm too
[https://github.com/jinyus/related_post_gen/pull/59](https://github.com/jinyus/related_post_gen/pull/59)