@fizmat 2023-10-02
The script does (a): `for p in heapq.nlargest(5, relation_count, key=relation_count.get)`
Internally the counter does (b): `for p, _n in heapq.nlargest(n, relation_count.items(), key=itemgetter(1))`
If I replace (a) with (b) in the script, it has the same performance as just calling relation_count.most_common(), which makes sense.

It seems (b) is measurably slower on almost all of my systems too. I wonder if it's because in (a) the heapq only stores the keys and gets counts from the counter, but in (b) the heapq stores the item-count pairs, so it uses 2x the memory.

Checked on a PC. (b) is slower than (a) on both Linux and Windows (both in python 3.11). Performance is the same *only* on M1 Mac in 3.11. I wonder how.

Win10 results (1: and 2: don't have the slower top5 from counter):
```
commit				import 		compile 	input 		process 		output 		total
0: install all pytho… 	11.90±0.18 	0.00±0 	10.42±0.26 	2744±18 	480.9±1.8 	3248±19
0: try counter most_… 	7.09±0.18 	0.00±0 	10.17±0.17 	2810±16 	54.28±0.99 	2882±16
1: minor pythonizati… 	12.08±0.28 	0.00±0 	10.27±0.26 	2729±13 	482.0±1.9 	3233±14
2: use ujson for IO…… 	7.12±0.23 	0.00±0 	10.21±0.17 	2718±14 	54.3±1.3 	2789±14
```
So this convenience change is definitely out.


By the way, on Windows the uint8 numpy matrix actually runs slower than uint16:
```
commit				import 		compile 	input 			process 		output 		total
0: minor pythonizati… 	115.92±0.72 	0.00±0 	10.35±0.21 		768.8±3.9 	481.6±1.5 	1376.7±4.0
1: use ujson for IO…… 	117.47±0.82 	0.00±0 	10.06±0.16 		770.8±3.4 	53.09±0.59 	951.4±4.0
2: np: use python se… 	118.21±0.91 	0.00±0 	10.031±0.098 	701.8±4.1 	52.74±0.63 	882.7±4.4
3: np: use smaller m… 	117.74±0.82 	0.00±0 	9.99±0.15 		743.1±2.1 	53.10±0.60 	923.9±2.7
4: np: use dict of t… 	118.3±1.3 	0.00±0 	10.03±0.14 		696.2±5.8 	53.15±0.61 	877.8±6.0
5: try to use numpy … 	117.56±0.87 	0.00±0 	9.95±0.17 		737.8±1.8 	53.03±0.68 	918.4±1.9
```

Is that a factor at all for you, or can performance differences on Windows be safely ignored?